Remove deprecated Integer boxing constructors
Mostly trivial like-for-like replacement, but for one instance in
IPAddressName where I removed the unnecessary creation of an integer and
instead created a Byte directly.

There was also some on-save Eclipse formatting, I would have liked that
to be a separate commit but seeing as how the commit is quite small I'm
allowing it.